# openapi.yaml
openapi: 3.1.0
info:
  title: Participant Registration API
  description: API to add participants to a Google Sheet
  version: 1.0.0
servers:
  - url: https://YOUR-RAILWAY-APP.up.railway.app
    description: Production server (replace with your Railway URL)
paths:
  /add-participant:
    post:
      operationId: addParticipant
      summary: Add a new participant to the registration sheet
      description: Adds a new row to the Google Sheet with participant information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - participantName
                - childDob
                - parentGuardianName
              properties:
                participantName:
                  type: string
                  description: The name of the participant (child)
                  example: "John Smith"
                childDob:
                  type: string
                  description: The child's date of birth
                  example: "2018-05-15"
                parentGuardianName:
                  type: string
                  description: The name of the parent or guardian
                  example: "Jane Smith"
      responses:
        "200":
          description: Participant added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  updatedRange:
                    type: string
        "400":
          description: Bad request - missing required fields
        "401":
          description: Unauthorized - invalid API key
        "500":
          description: Server error
  /participants:
    get:
      operationId: getParticipants
      summary: Get all registered participants
      description: Retrieves all participants from the Google Sheet
      responses:
        "200":
          description: List of participants
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  participants:
                    type: array
                    items:
                      type: object
                      properties:
                        participantName:
                          type: string
                        childDob:
                          type: string
                        parentGuardianName:
                          type: string
        "401":
          description: Unauthorized - invalid API key
        "500":
          description: Server error